requirements.txt
attrs==24.2.0
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
flatbuffers==24.3.25
idna==3.7
numpy==2.0.1
openmeteo_requests==1.3.0
openmeteo_sdk==1.14.1
pandas==2.2.2
platformdirs==4.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
six==1.16.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.2
